feat: carry a project's color on the glyphs its rows already draw
Review feedback on #341. A stripe per tab row is a second mark on a row
that already has one, and it indented tagged rows 9pt past untagged ones.
The tag now tints the icon each row already draws — the project's own and
each of its tabs' — so a tagged sidebar occupies exactly the space an
untagged one does. The stripe stays only for the "None" icon preference,
the one case with no glyph to tint, on both row kinds. An agent logo keeps
its brand color: it is a live status signal, and a tab's project is
already named by the section above it.

The titlebar badge drew the project's own icon, which a toolbar item lays
out at control size — outsizing the title beside it. It draws a plain dot
at an explicit diameter instead. That also drops the only cross-file user
of SidebarRowIcon, so it moves back into Sidebar.swift.
