+++ AGENTS.md
@@ -196,9 +196,11 @@ The seed is also **off under `MACTERM_BENCHMARK=1`** (`BenchmarkControl.isEnable
 
 ### Project Color Tags
 
-A project can carry an optional color tag (`Project.colorName`, set from the sidebar's project context menu or Settings → Projects). It renders in three places: the sidebar folder icon, a leading stripe on each of that project's tab rows (`ProjectColorStripe` — the project's own row has no stripe, its leading edge is the disclosure triangle's), and beside the window title (`ProjectColorBadge`), which draws the project's OWN icon in that color rather than an abstract dot. It renders through the same `SidebarRowIcon` the row uses (extracted out of `Sidebar.swift` for exactly this) — deriving the symbol independently was tried first and disagreed with the sidebar the moment the icon preference moved off `folder`, putting a numbered row beside a folder in the titlebar. "None" is the one choice that can't be honored literally, since the badge would vanish and take the color with it; it falls back to a plain dot, mirroring the row, which shows its color as a stripe when it has no glyph.
+A project can carry an optional color tag (`Project.colorName`, set from the sidebar's project context menu or Settings → Projects). **It tints a glyph the row already draws** — the project's own icon, and the icon on each of that project's tab rows (`SidebarTabRow.iconTint`, threaded on through `TabStatusGlyph` so a running or done row keeps its color) — so a tagged sidebar occupies exactly the space an untagged one does. A leading stripe per tab row was the first shape this took and was rejected in review: it is a second mark where the row already has one, and it indented tagged rows 9pt past untagged ones. The stripe survives (`ProjectColorStripe`) as the fallback for the one case with no glyph to tint — the "None" icon preference — on both row kinds. An agent logo never wears the tag: `SidebarRowIcon` paints its brand color from the inside, and that logo is a live status signal, while the project a tab belongs to is already named by the section above it.
 
-Three things are deliberate. The colors are **fixed system colors** (`MactermTheme.color(for:)`), the same deliberate exception to "all colors come from MactermTheme" that `AgentIcon.brandColor` takes — these are identity labels the user picked, not chrome that should track the terminal. Deriving them from the user's ghostty palette was tried first and is a trap: a theme puts any color in any ANSI slot (under Rose Pine, Macterm's own default, slot 6 "cyan" is a rosy beige and slot 2 "green" is a teal blue), and Rose Pine additionally aliases its bright slots onto its base ones (`9 == 1`, `13 == 5`), which collapsed eight tags into six identical-looking ones. The test that guarded this originally compared slot *indices* and so proved nothing; it now compares resolved sRGB components. The picker is `.pickerStyle(.palette)` (`ProjectColorMenu`, shared by the sidebar context menu and Settings → Projects) — Apple's control for a row of colors inside a menu, and the only thing that works: a `Label` whose icon is a tinted SF Symbol renders in the MENU's tint (AppKit treats a menu item's symbol as a template), and one whose icon is an explicit non-template `NSImage` is dropped by SwiftUI's menu bridge entirely. Both were tried and observed failing in the running app; don't re-attempt either, and note the palette's selection ring is what shows the current color, since a swatch on the "Color" row itself is subject to the same limitation. The stored value is a **String, not the enum**, so a tag written by a newer build (or a hand-edited `projects.json`) degrades to "untagged" instead of failing the whole row's decode. And a stripe is drawn only for a tagged project — no gutter is reserved — so an untagged sidebar sits exactly where it always did, at the cost of tagged rows indenting 9pt further than untagged ones.
+Beside the window title the tag shows as a plain dot (`ProjectColorBadge`), which is the one place it is NOT the row's own glyph. Drawing the project's icon there was tried first — it kept a numbered sidebar row from disagreeing with a folder in the titlebar — and was dropped in review: at titlebar size a tinted folder reads as a control rather than a tag, and a toolbar item lays its content out at control size, so the icon outsized the title beside it. The dot is a `Circle` at an explicit `@ScaledMetric` diameter for exactly that reason; it renders for a tagged project only, and never for the pinned rows, which have no project.
+
+Three things are deliberate. The colors are **fixed system colors** (`MactermTheme.color(for:)`), the same deliberate exception to "all colors come from MactermTheme" that `AgentIcon.brandColor` takes — these are identity labels the user picked, not chrome that should track the terminal. Deriving them from the user's ghostty palette was tried first and is a trap: a theme puts any color in any ANSI slot (under Rose Pine, Macterm's own default, slot 6 "cyan" is a rosy beige and slot 2 "green" is a teal blue), and Rose Pine additionally aliases its bright slots onto its base ones (`9 == 1`, `13 == 5`), which collapsed eight tags into six identical-looking ones. The test that guarded this originally compared slot *indices* and so proved nothing; it now compares resolved sRGB components. The picker is `.pickerStyle(.palette)` (`ProjectColorMenu`, shared by the sidebar context menu and Settings → Projects) — Apple's control for a row of colors inside a menu, and the only thing that works: a `Label` whose icon is a tinted SF Symbol renders in the MENU's tint (AppKit treats a menu item's symbol as a template), and one whose icon is an explicit non-template `NSImage` is dropped by SwiftUI's menu bridge entirely. Both were tried and observed failing in the running app; don't re-attempt either, and note the palette's selection ring is what shows the current color, since a swatch on the "Color" row itself is subject to the same limitation. The stored value is a **String, not the enum**, so a tag written by a newer build (or a hand-edited `projects.json`) degrades to "untagged" instead of failing the whole row's decode. And the tint is applied only for a tagged project: a tab row's icon was already forced to `.secondary`, so the tag is a straight swap there, but the project row's icon carries no style of its own and forcing one on it would flatten the label AppKit gives a selected row.
 
 A new project can be tagged automatically (`Preferences.autoAssignProjectColors`, Settings → Appearance, **default off**): `ProjectStore.create` asks `ProjectColor.leastUsed(among:)`, which is one rule rather than two — with every color used equally the tie-break returns to the front of `allCases`, so the first eight projects get eight distinct colors and only the ninth repeats. The preference is read at creation ONLY: flipping it never tags existing projects nor clears tags already set. `ProjectStore` reads it through an injectable closure (`autoAssignsColors`) because swift-testing runs suites in parallel and a test mutating the real preference would race every other one.
